Tests/LibWeb/Text/input/CredentialManagement/PasswordCredential-ctors.html

<!DOCTYPE html>
<script src="../include.js"></script>
<form id="loginForm">
    <input type="text" name="username" autocomplete="username" value="staplemachine@example.com" required>
    <input type="password" name="password" autocomplete="current-password" value="correctstaplehorsecinema" required>
    <input type="text" name="name" autocomplete="name" value="Staple Machine" required>
    <input type="text" name="photo" autocomplete="photo" value="http://example.com/images/horse.png" required>
</form>
<form id="changePasswordForm">
    <input type="text" name="username" autocomplete="username" value="staplemachine@example.com" required>
    <input type="password" name="password" autocomplete="new-password" value="correctstaplehorsecinema" required>
    <input type="password" name="password" autocomplete="current-password" value="correcthorsebatterystaple" required>
</form>
<form id="loginFormWithNick">
    <input type="text" name="username" autocomplete="username" value="staplemachine@example.com" required>
    <input type="password" name="password" autocomplete="current-password" value="correcthorsebatterystaple" required>
    <input type="text" name="name" autocomplete="nickname" value="StapleNick" required>
    <input type="text" name="photo" autocomplete="photo" value="http://example.com/images/horse.png" required>
</form>
<script>
test(() => {
    const data = {
        id: 'staplemachine@example.com',
        password: 'correcthorsebatterystaple',
        name: 'Staple Machine',
        iconURL: 'http://example.com/images/horse.png',
    };
    let cred = new PasswordCredential(data);
    println(`cred instanceof PasswordCredential: ${cred instanceof PasswordCredential}`)
    println(`id: ${cred.id}`, cred.id);
    println(`password: ${cred.password}`, cred.password);
    println(`name: ${cred.name}`, cred.name);
    println(`iconURL: ${cred.iconURL}`, cred.iconURL);

    let form = document.getElementById('loginForm');
    cred = new PasswordCredential(form);
    println(`cred instanceof PasswordCredential: ${cred instanceof PasswordCredential}`)
    println(`id: ${cred.id}`, cred.id);
    println(`password: ${cred.password}`, cred.password);
    println(`name: ${cred.name}`, cred.name);
    println(`iconURL: ${cred.iconURL}`, cred.iconURL);

    form = document.getElementById('changePasswordForm');
    cred = new PasswordCredential(form);
    println(`cred instanceof PasswordCredential: ${cred instanceof PasswordCredential}`)
    println(`id: ${cred.id}`, cred.id);
    println(`password: ${cred.password}`, cred.password);

    form = document.getElementById('loginFormWithNick');
    cred = new PasswordCredential(form);
    println(`cred instanceof PasswordCredential: ${cred instanceof PasswordCredential}`)
    println(`id: ${cred.id}`, cred.id);
    println(`password: ${cred.password}`, cred.password);
    println(`name: ${cred.name}`, cred.name);
    println(`iconURL: ${cred.iconURL}`, cred.iconURL);
});
</script>
