Tests/LibWeb/Text/expected/CredentialManagement/PasswordCredential-ctors.txt

cred instanceof PasswordCredential: true
id: staplemachine@example.com
password: correcthorsebatterystaple
name: Staple Machine
iconURL: http://example.com/images/horse.png
cred instanceof PasswordCredential: true
id: staplemachine@example.com
password: correctstaplehorsecinema
name: Staple Machine
iconURL: http://example.com/images/horse.png
cred instanceof PasswordCredential: true
id: staplemachine@example.com
password: correctstaplehorsecinema
cred instanceof PasswordCredential: true
id: staplemachine@example.com
password: correcthorsebatterystaple
name: StapleNick
iconURL: http://example.com/images/horse.png
